Clearer style for open projects
The blue background was easily mistaken to mean that the item was
open in the fuzzy finder. Now, we bold the text and use a different
icon (an open book instead of a closed book!).

Unrelatedly, added reset-project to package.json, so that atom knows
about the command during startup.